Generate unique ID and quantity for each leaf item selected
Rendered components recursively, with each one getting an ID which is prefixed by it's parents. When a leaf node is selected, the compounded ID is passed to a top level callback, along with the selected quanity.

This makes the components themselves a lot cleaner, and all of the data needed to manage the selected items is in a simple key-value pair. This should allow the ability to offload the work of validating the selection to a util module. It should be able to dissect the keys, and build up the required metadata (e.g. total price of selected, selection limits etc) using the reference data from the server.